requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
drf-spectacular==0.27.2

psycopg2-binary
python-dotenv==1.0.1

sqlparse==0.5.2
tzdata==2024.2

PyYAML==6.0.2
PyJWT==2.10.1



